+++ packages/contracts-rfq/contracts/interfaces/IFastBridgeV2.sol
@@ -15,7 +15,7 @@ interface IFastBridgeV2 is IFastBridge {
     struct BridgeTxDetails {
         BridgeStatus status;
         uint32 destChainId;
-        uint40 proofBlockTimestamp;
+        uint56 proofBlockTimestamp;
         address proofRelayer;
     }
